back/requirements.txt
python-docx==1.1.0
requests==2.32.0
google-auth==2.28.1
google-cloud-texttospeech==2.16.2
pydub==0.25.1
pika==1.3.2
Flask==3.0.2
gunicorn==21.2.0
kubernetes==29.0.0
python-dotenv==1.0.1  # para cargar tokens desde .env
pydantic==2.10.0      # opcional, si quieres validar respuestas con modelos
loguru==0.7.2         # para logs bonitos (opcional)